python-webparse/requirements.txt
feedparser
pandas
requests
bs4
lxml
newspaper3k
lxml_html_clean
